phanns/pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel", "Cython >= 3.0.0", "numpy >= 1.26.0"]
build-backend = "setuptools.build_meta"

[project]
name = "phanns"
version = "0.0.1"
description = "Annotate phage structural proteins using a neural network."
dependencies = [
    "biopython>=1.80",
    "numpy>=1.26.0",
    "tensorflow==2.16.1",
    "pyarrow>=15.0.0",
    "tqdm>=4.66.0",
    "scikit-learn>=1.4.0",
    ]
requires-python = ">= 3.8"
authors = [
  {name = "Sean Fahey", email = "seanfahey21@gmail.com"},
]
license = {text = "MIT License"}
classifiers = [
  "Development Status :: 3 - Alpha",

  "Intended Audience :: Education",
  "Topic :: Scientific/Engineering :: Bio-Informatics",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",

  "Operating System :: OS Independent",
]

[project.urls]
Repository = "https://github.com/seanfahey1/PhANNs3"

[project.scripts]
phanns = "phanns:main"
